models, brands and manufacturers now implemented in a basic way
Manufacturers page is passed all models it produced in order to create a
list of models.  Intention is to have a selector to slim down by brand.
Also noticed that John Stokes' books don't have Clipper listed as a
brand - a glaring omission?  This was a major brand of cheap models for
Akrad towards the end of the valve era.